Move TreeMapWithStages to dotty.tools.dotc.staging
